+++ FluentFTP/Proxy/SyncProxy/FtpClientBlueCoatProxy.cs
@@ -1,20 +1,25 @@
 using FluentFTP.Client.BaseClient;
 
-namespace FluentFTP.Proxy {
-	/// <summary> A FTP client with a user@host proxy identification. </summary>
-	public class FtpClientUserAtHostProxy : FtpClientProxy {
+namespace FluentFTP.Proxy.SyncProxy {
+	/// <summary> 
+	/// A FTP client with a user@host proxy identification, that works with Blue Coat FTP Service servers.
+	/// 
+	/// The 'blue coat variant' forces the client to wait for a 220 FTP response code in 
+	/// the handshake phase.
+	/// </summary>
+	public class FtpClientBlueCoatProxy : FtpClientProxy {
 		/// <summary> A FTP client with a user@host proxy identification. </summary>
 		/// <param name="proxy">Proxy information</param>
-		public FtpClientUserAtHostProxy(FtpProxyProfile proxy)
+		public FtpClientBlueCoatProxy(FtpProxyProfile proxy)
 			: base(proxy) {
-			ConnectionType = "User@Host";
+			ConnectionType = "Blue Coat";
 		}
 
 		/// <summary>
 		/// Creates a new instance of this class. Useful in FTP proxy classes.
 		/// </summary>
 		protected override BaseFtpClient Create() {
-			return new FtpClientUserAtHostProxy(Proxy);
+			return new FtpClientBlueCoatProxy(Proxy);
 		}
 
 		/// <summary> Redefine the first dialog: auth with proxy information </summary>
@@ -26,6 +31,14 @@ namespace FluentFTP.Proxy {
 
 			// Connection USER@Host means to change user name to add host.
 			Credentials.UserName = Credentials.UserName + "@" + Host + ":" + Port;
+
+			var reply = GetReply();
+			if (reply.Code == "220") {
+				LogLine(FtpTraceLevel.Info, "Status: Server is ready for the new client");
+			}
+
+			// TO TEST: if we are able to detect the actual FTP server software from this reply
+			HandshakeReply = reply;
 		}
 	}
 }