+++ FluentFTP/Proxy/SyncProxy/FtpClientHttp11Proxy.cs
@@ -6,7 +6,7 @@ using FluentFTP.Client.BaseClient;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace FluentFTP.Proxy {
+namespace FluentFTP.Proxy.SyncProxy {
 	/// <summary> A FTP client with a HTTP 1.1 proxy implementation. </summary>
 	public class FtpClientHttp11Proxy : FtpClientProxy {
 		/// <summary> A FTP client with a HTTP 1.1 proxy implementation </summary>
@@ -43,16 +43,6 @@ namespace FluentFTP.Proxy {
 			Connect(stream, Host, Port, FtpIpVersion.ANY);
 		}
 
-#if ASYNC
-		/// <summary>
-		/// Connects to the server using an existing <see cref="FtpSocketStream"/>
-		/// </summary>
-		/// <param name="stream">The existing socket stream</param>
-		protected override Task ConnectAsync(FtpSocketStream stream, CancellationToken token) {
-			return ConnectAsync(stream, Host, Port, FtpIpVersion.ANY, token);
-		}
-#endif
-
 		/// <summary>
 		/// Connects to the server using an existing <see cref="FtpSocketStream"/>
 		/// </summary>
@@ -78,49 +68,12 @@ namespace FluentFTP.Proxy {
 			ProxyHandshake(stream);
 		}
 
-#if ASYNC
-		/// <summary>
-		/// Connects to the server using an existing <see cref="FtpSocketStream"/>
-		/// </summary>
-		/// <param name="stream">The existing socket stream</param>
-		/// <param name="host">Host name</param>
-		/// <param name="port">Port number</param>
-		/// <param name="ipVersions">IP version to use</param>
-		/// <param name="token">IP version to use</param>
-		protected override async Task ConnectAsync(FtpSocketStream stream, string host, int port, FtpIpVersion ipVersions, CancellationToken token) {
-			await base.ConnectAsync(stream, token);
-
-			var writer = new StreamWriter(stream);
-			await writer.WriteLineAsync(string.Format("CONNECT {0}:{1} HTTP/1.1", host, port));
-			await writer.WriteLineAsync(string.Format("Host: {0}:{1}", host, port));
-			if (Proxy.ProxyCredentials != null) {
-				var credentialsHash = Convert.ToBase64String(System.Text.Encoding.UTF8.GetBytes(Proxy.ProxyCredentials.UserName + ":" + Proxy.ProxyCredentials.Password));
-				await writer.WriteLineAsync("Proxy-Authorization: Basic " + credentialsHash);
-			}
-
-			await writer.WriteLineAsync("User-Agent: custom-ftp-client");
-			await writer.WriteLineAsync();
-			await writer.FlushAsync();
-
-			await ProxyHandshakeAsync(stream, token);
-		}
-#endif
-
 		private void ProxyHandshake(FtpSocketStream stream) {
 			var proxyConnectionReply = GetProxyReply(stream);
 			if (!proxyConnectionReply.Success) {
 				throw new FtpException("Can't connect " + Host + " via proxy " + Proxy.ProxyHost + ".\nMessage : " + proxyConnectionReply.ErrorMessage);
 			}
 		}
-
-#if ASYNC
-		private async Task ProxyHandshakeAsync(FtpSocketStream stream, CancellationToken token = default(CancellationToken)) {
-			var proxyConnectionReply = await GetProxyReplyAsync(stream, token);
-			if (!proxyConnectionReply.Success) {
-				throw new FtpException("Can't connect " + Host + " via proxy " + Proxy.ProxyHost + ".\nMessage : " + proxyConnectionReply.ErrorMessage);
-			}
-		}
-#endif
 
 		private FtpReply GetProxyReply(FtpSocketStream stream) {
 			var reply = new FtpReply();
@@ -162,44 +115,5 @@ namespace FluentFTP.Proxy {
 			return reply;
 		}
 
-#if ASYNC
-		private async Task<FtpReply> GetProxyReplyAsync(FtpSocketStream stream, CancellationToken token = default(CancellationToken)) {
-			var reply = new FtpReply();
-			string buf;
-
-			if (!IsConnected) {
-				throw new InvalidOperationException("No connection to the server has been established.");
-			}
-
-			stream.ReadTimeout = ReadTimeout;
-			while ((buf = await stream.ReadLineAsync(Encoding, token)) != null) {
-				Match m;
-
-				LogLine(FtpTraceLevel.Info, buf);
-
-				if ((m = Regex.Match(buf, @"^HTTP/.*\s(?<code>[0-9]{3}) (?<message>.*)$")).Success) {
-					reply.Code = m.Groups["code"].Value;
-					reply.Message = m.Groups["message"].Value;
-					break;
-				}
-
-				reply.InfoMessages += buf + "\n";
-			}
-
-			// fixes #84 (missing bytes when downloading/uploading files through proxy)
-			while ((buf = await stream.ReadLineAsync(Encoding, token)) != null) {
-				LogLine(FtpTraceLevel.Info, buf);
-
-				if (Strings.IsNullOrWhiteSpace(buf)) {
-					break;
-				}
-
-				reply.InfoMessages += buf + "\n";
-			}
-
-			return reply;
-		}
-
-#endif
 	}
 }