+++ FluentFTP/Proxy/SyncProxy/FtpClientSocks4Proxy.cs
@@ -1,12 +1,38 @@
-﻿using FluentFTP.Proxy.Socks;
+﻿
+/* Unmerged change from project 'FluentFTP (net472)'
+Before:
+using FluentFTP.Proxy.Socks;
+After:
+using FluentFTP;
+using FluentFTP.Proxy;
+using FluentFTP.Proxy.Socks;
+*/
+
+/* Unmerged change from project 'FluentFTP (net462)'
+Before:
+using FluentFTP.Proxy.Socks;
+After:
+using FluentFTP;
+using FluentFTP.Proxy;
+using FluentFTP.Proxy.Socks;
+*/
+using FluentFTP.Client.BaseClient;
+using FluentFTP.Proxy.Socks;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace FluentFTP.Proxy {
+namespace FluentFTP.Proxy.SyncProxy {
 	/// <summary> A FTP client with a SOCKS4 proxy implementation. </summary>
 	public class FtpClientSocks4Proxy : FtpClientProxy {
 		public FtpClientSocks4Proxy(FtpProxyProfile proxy) : base(proxy) {
 			ConnectionType = "SOCKS4 Proxy";
+		}
+
+		/// <summary>
+		/// Creates a new instance of this class. Useful in FTP proxy classes.
+		/// </summary>
+		protected override BaseFtpClient Create() {
+			return new FtpClientSocks4Proxy(Proxy);
 		}
 
 		/// <summary>
@@ -16,16 +42,6 @@ namespace FluentFTP.Proxy {
 			BaseStream.Read(new byte[6], 0, 6);
 			base.Handshake();
 		}
-
-#if ASYNC
-		/// <summary>
-		/// Called during <see cref="ConnectAsync()"/>. Typically extended by FTP proxies.
-		/// </summary>
-		protected virtual async Task HandshakeAsync(CancellationToken token = default(CancellationToken)) {
-			await BaseStream.ReadAsync(new byte[6], 0, 6);
-			base.Handshake();
-		}
-#endif
 
 		protected override void Connect(FtpSocketStream stream) {
 			base.Connect(stream);
@@ -39,12 +55,5 @@ namespace FluentFTP.Proxy {
 			proxy.Connect();
 		}
 
-#if ASYNC
-		protected override async Task ConnectAsync(FtpSocketStream stream, CancellationToken cancellationToken) {
-			await base.ConnectAsync(stream, cancellationToken);
-			var proxy = new Socks4Proxy(Host, Port, stream);
-			await proxy.ConnectAsync();
-		}
-#endif
 	}
 }