+++ FluentFTP/Proxy/SyncProxy/FtpClientSocks4aProxy.cs
@@ -3,11 +3,18 @@ using FluentFTP.Client.BaseClient;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace FluentFTP.Proxy {
+namespace FluentFTP.Proxy.SyncProxy {
 	/// <summary> A FTP client with a SOCKS4a proxy implementation. </summary>
 	public class FtpClientSocks4aProxy : FtpClientProxy {
 		public FtpClientSocks4aProxy(FtpProxyProfile proxy) : base(proxy) {
 			ConnectionType = "SOCKS4a Proxy";
+		}
+
+		/// <summary>
+		/// Creates a new instance of this class. Useful in FTP proxy classes.
+		/// </summary>
+		protected override BaseFtpClient Create() {
+			return new FtpClientSocks4aProxy(Proxy);
 		}
 
 		/// <summary>
@@ -17,16 +24,6 @@ namespace FluentFTP.Proxy {
 			BaseStream.Read(new byte[6], 0, 6);
 			base.Handshake();
 		}
-
-#if ASYNC
-		/// <summary>
-		/// Called during <see cref="ConnectAsync()"/>. Typically extended by FTP proxies.
-		/// </summary>
-		protected virtual async Task HandshakeAsync(CancellationToken token = default(CancellationToken)) {
-			await BaseStream.ReadAsync(new byte[6], 0, 6);
-			base.Handshake();
-		}
-#endif
 
 		protected override void Connect(FtpSocketStream stream) {
 			base.Connect(stream);
@@ -40,12 +37,5 @@ namespace FluentFTP.Proxy {
 			proxy.Connect();
 		}
 
-#if ASYNC
-		protected override async Task ConnectAsync(FtpSocketStream stream, CancellationToken cancellationToken) {
-			await base.ConnectAsync(stream, cancellationToken);
-			var proxy = new Socks4aProxy(Host, Port, stream);
-			await proxy.ConnectAsync();
-		}
-#endif
 	}
 }